pyproject.toml
[project]
name = "handwritten-ocr"
version = "0.1.0"
description = "Local-first OCR pipeline for converting handwritten notes to text using vision-language models."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    # --- Core (directly imported) ---
    "torch",
    "transformers>=4.40.0",
    "pillow",
    "pydantic>=2.0",
    "ollama>=0.4",
    "opencv-python",
    "numpy",
    "langgraph (>=1.0.8,<2.0.0) ; python_version >= \"3.10\" and python_version < \"4.0\"",
    # --- Transitive (required at runtime by transformers / olmOCR, not directly imported) ---
    "torchvision",
    "sentencepiece",
    "accelerate",
    "tiktoken",
    "verovio",
    "compressed-tensors",
    "python-dotenv",
    # --- rag tooling layer ---
    "qdrant-client>=1.9",
    "fastembed>=0.3",
    "onnxruntime>=1.17,<1.24",
    "open-clip-torch>=2.24",
    "loguru>=0.7",
]

[project.optional-dependencies]
notebooks = ["jupyter", "ipykernel", "matplotlib"]

[project.scripts]
ocr = "ocr_agent.transcribe:main"

[tool.poetry]
packages = [{ include = "ocr_agent" }, { include = "rag" }]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
